Add collection of drugs and consumables

+++ DarkHeresy2CharacterCreator/DarkHeresy2CharacterCreator/Model/Equipment/Interfaces/IItem.cs
@@ -12,7 +12,7 @@ namespace DarkHeresy2CharacterCreator.Model.Equipment.Interfaces
         string Name { get; set; }
         bool IsEquiped { get; set; }
         Availability Availability { get; }
-        int Weight { get; }
+        double Weight { get; }
         string Discription { get; }
         Craftsmanship Craftsmanship { get; set; }
     }
